Place shaders in their own header files. Remove shader files from CMake GLBUILD_SRCS. Add Unofficial libsquish from vcpkg; removing the need for internal libsquish (to be removed). Remove libsquish sources from OPENGL_ENGINE_BUILD_SRCS in CMake. Add target_compile_options for build_engine and kenbuild for OpenGL / PolyMost. Remove KENBUILD_GAME_DATA list from CMake. Use <numbers> for pi. Remove extern from kplib.hpp functions. Fix const char* -> char* cast in mdsprite.cpp. Remove need for MAXWALLSB variable usage in polymost_priv.hpp. Fix cast in kprender in PTM_LoadTextureFile. Include "polymost_priv.hpp" in polymosttexcompress.cc (for gltexcomprquality).

+++ src/polymosttex.cpp
@@ -356,7 +356,7 @@ int PTM_LoadTextureFile(const char* filename, PTMHead* ptmh, int flags, int effe
 	}
 	memset(tex.pic, 0, tex.sizx * tex.sizy * sizeof(coltype));
 
-	if (kprender(picdata, picdatalen, tex.pic, tex.sizx * sizeof(coltype), tex.sizx, tex.sizy, 0, 0)) {
+	if (kprender(picdata, picdatalen, (intptr_t)tex.pic, tex.sizx * sizeof(coltype), tex.sizx, tex.sizy, 0, 0)) {
 		free(picdata);
 		free(tex.pic);
 		return -5;
